fix some type conversion warnings
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@4133 599ab11a-0111-0410-8a50-d7a5b0953570

+++ core/dccm/synce-device-legacy.c
@@ -86,7 +86,7 @@ synce_device_legacy_send_ping(gpointer data)
 }
 
 static void
-synce_device_legacy_info_received (SynceDeviceLegacy *self, const guchar *buf, gint length)
+synce_device_legacy_info_received (SynceDeviceLegacy *self, const guchar *buf, gssize length)
 {
   SynceDevicePrivate *priv = SYNCE_DEVICE_GET_PRIVATE (SYNCE_DEVICE(self));
